+++ src/core/utils/mod.rs
@@ -1,3 +1,3 @@
 pub mod defaults {
-    pub(crate) const DEFAULT_FPS: u8 = 24;
+    pub(crate) const DEFAULT_FPS: u32 = 24;
 }
